fix close behavior
There was a network race condition where the socket was TCP RST
immediately after the TCP FIN, before the 0 had been returned from the
read. One way to handle this is to send some other additional data to
guarentee a non-zero read, or to just ignore the close error. We choose
the latter to keep things nice and simple.